requirements.txt
Flask
flask-cors
flask[async]
python-telegram-bot==20.3
python-telegram-bot[passport,urllib3,socks]
python-dotenv
google-generativeai
pandas
openpyxl
matplotlib
urllib3
six
